Adding js validation for the register form. To do that I have used the jquery validation engine and the jquery i18n plugin to internationalize it. It's now easy to add validation of any kind to other form and to translate other part in js.
The validation is divided in different part.
All the validations that are written above the fields
are enforced like only number & letter or min
6 char.
But there is also an ajax call that's made to make
sure the username is not already taken.
If the form fail to comply with any of those validation
it won't be send and a prompt will show with a
message explaining the problem.