Fix all problematic basename calls in classes
basename will strip the first character if it is accented.
This commit add the mb_basename function, taken from https://bugs.php.net/bug.php?id=37268

This fixes:
- file renaming on upload when a file with the same name already exists
- comments and permissions ("éa.jpg" and "a.jpg" used the same file for comments/permissions)
- photospheres folder ("éPhotoSpheres" was matching "PhotoSpheres")
- file and folder name on the right menu
- file and folder name in the rename textbox
- filename in the left menu, the header and the RSS file
- filename when downloading a picture or a zip

The only remaining problem is with the exif infos on the right menu, but it is an external call to exif_read_data().