CSS: display #image_panel elements as table-rows
There is now a new div (#image_panel_table) inside #image_panel, encompassing
#bigimage and #image_bar.
This change was made to have dynamic resizing inside #image_panel, which
will help to display description (see next commit) with a variable height.

It should not change the page rendering, with the exception of a picture
displayed with a page height under 300px. Instead of reducing more and more
#image_panel, now a scrollbar appears.