New feature: add a description to a folder or a picture
This new feature is optional (but activated by default).
A new option added to the right menu permits to write a description, displayed
between the image/video and the navigation buttons, or under the header for a
folder.
When a description is empty, it is hidden using CSS to remove its padding.

The Description class is based on the Comments class.